+++ helloboot/src/main/java/tobysrping/helloboot/HellobootApplication.java
@@ -1,28 +1,10 @@
 package tobysrping.helloboot;
 
 import org.springframework.boot.SpringApplication;
-import org.springframework.boot.web.embedded.tomcat.TomcatServletWebServerFactory;
-import org.springframework.boot.web.servlet.server.ServletWebServerFactory;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.servlet.DispatcherServlet;
 
-@Configuration
-@ComponentScan
+@MySpringBootAnnotation
 public class HellobootApplication {
-	@Bean
-	public ServletWebServerFactory serverWebServerFactory() {
-		return new TomcatServletWebServerFactory();
-	}
-
-	@Bean
-	public DispatcherServlet dispatcherServlet() {
-		return new DispatcherServlet();
-	}
-
 	public static void main(String[] args) {
 		SpringApplication.run(HellobootApplication.class, args);
 	}
-
 }
